NTV-9 Delete the old pure reactor way of doing things, refactor traditional servlet api to be the only way.
