--- src/main/kotlin/com/ngenenius/api/reactive/controller/TwitchController.kt
@@ -1,24 +1,0 @@
-package com.ngenenius.api.reactive.controller
-
-import com.ngenenius.api.model.twitch.StreamDetailsResponse
-import com.ngenenius.api.reactive.service.TwitchService
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.RestController
-import reactor.core.publisher.Mono
-
-@Deprecated("Will be removed")
-@RestController
-@RequestMapping("/twitch")
-class TwitchController(private val twitchService: TwitchService) {
-
-    @GetMapping("/team-view")
-    fun teamViewStreamDetails(): Mono<StreamDetailsResponse> {
-        return twitchService.teamViewStreamDetails()
-    }
-
-    @GetMapping("/tournament")
-    fun tournamentStreamDetails(): Mono<StreamDetailsResponse> {
-        return twitchService.tournamentViewStreamDetails()
-    }
-}
