+++ src/test/kotlin/com/ngenenius/api/controller/TwitchControllerTest.kt
@@ -1,16 +1,16 @@
-package com.ngenenius.api.web.controller
+package com.ngenenius.api.controller
 
-import com.ngenenius.api.web.service.TwitchServiceV2
+import com.ngenenius.api.service.TwitchService
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
 import org.junit.jupiter.api.Test
 
-internal class TwitchControllerV2Test {
+internal class TwitchControllerTest {
 
-    private val twitchService = mock<TwitchServiceV2>()
-    private val controller = TwitchControllerV2(twitchService)
+    private val twitchService = mock<TwitchService>()
+    private val controller = TwitchController(twitchService)
 
     @Test
     fun `Should use the twitch service for team view details when getting team view`() {
